LayoutTests/http/tests/security/xssAuditor/malformed-xss-protection-header-9-expected.txt

CONSOLE MESSAGE: line 1: Error parsing header X-XSS-Protection: 1; mode=block; report=/fail; report=/fail;: duplicate report directive at character position 35. The default protections will be applied.
CONSOLE MESSAGE: line 4: Refused to execute a JavaScript script. Source code of script found within request.

This tests that the X-XSS-Protection header is not ignored when there is a duplicate report directive, and we issue an error



--------
Frame: 'frame'
--------
If you see this message and no JavaScript alert() then the test PASSED.
